requirements.txt
flask
flask-cors
sqlalchemy
bcrypt
requests
langchain
google-generativeai
openai
pdfplumber